Create smb.conf
After hundreds of hours patching broken samba and over several years debugging issues with windows, mac & linux (windows being the culprit ironically) and googling. This is a /etc/samba/smb.conf file that works well for me on a home network.  Probably not best for anywhere else but a local network